+++ src/Fable.Transforms/Python/Fable2Python.Types.fs
@@ -42,6 +42,7 @@ type DecoratorInfo =
     {
         Decorator: string
         Parameters: string
+        ImportFrom: string
     }